Sorts and adds XPath related tests
